Show entire workflow of the program
