+++ bluetooth/main.py
@@ -1,21 +1,20 @@
 import wifi
-import helper
+import logger
+import bluetoothManager
 
 def isConnectedTOWifi():
     return wifi.ConnectedToTheNetwork()
 
 def establishConnection():
-    helper.log("Trying to establish a wifi connection over bluetooth")
+    logger.log("Trying to establish a wifi connection over bluetooth")
 
 def start_farm():
-    helper.log("Starting farm", helper.LOG_DEBUG)
-    helper.log("Starting farm", helper.LOG_WARM)
-    helper.log("Starting farm", helper.LOG_ERROR)
+    logger.log("Starting farm", logger.LOG_DEBUG)
 
 
 if __name__ == "__main__":
     if isConnectedTOWifi():
-        helper.log("Connected to the wifi")
+        logger.log("Connected to the wifi")
         start_farm()
     establishConnection()
     start_farm()
